pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "barstar-backend"
version = "0.1.0"
description = "Barstar backend service."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Barstar Team" },
]
dependencies = [
  "alembic>=1.13",
  "fastapi>=0.111",
  "orjson>=3.10",
  "psycopg[binary]>=3.1",
  "pydantic-settings>=2.2",
  "python-dotenv>=1.0",
  "redis>=5.0",
  "sqlalchemy>=2.0",
  "structlog>=24.1",
  "uvicorn[standard]>=0.29",
]

[project.optional-dependencies]
dev = [
  "black>=24.4",
  "httpx>=0.27",
  "mypy>=1.9",
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "ruff>=0.5",
  "types-redis>=4.6",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
addopts = "-ra --strict-config --strict-markers"
asyncio_mode = "auto"
filterwarnings = [
  "error",
  "ignore::DeprecationWarning:pkg_resources",
]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["app", "tests"]
select = [
  "E",
  "F",
  "I",
  "B",
  "Q",
  "UP",
  "RUF",
]
unfixable = ["B905"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = "\\.pyi?$"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unreachable = true
disallow_any_unimported = false
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true
